refactor: separate parameters classes into different files #608
