+++ Example/PrebidDemoJava/src/main/java/org/prebid/mobile/javademo/activities/ads/gam/original/GamOriginalApiVideoRewarded.java
@@ -13,7 +13,7 @@ import com.google.android.gms.ads.rewarded.RewardedAdLoadCallback;
 
 import org.prebid.mobile.RewardedVideoAdUnit;
 import org.prebid.mobile.Signals;
-import org.prebid.mobile.VideoBaseAdUnit;
+import org.prebid.mobile.VideoParameters;
 import org.prebid.mobile.javademo.activities.BaseAdActivity;
 
 import java.util.Collections;
@@ -35,7 +35,7 @@ public class GamOriginalApiVideoRewarded extends BaseAdActivity {
     private void createAd() {
         adUnit = new RewardedVideoAdUnit(CONFIG_ID);
 
-        VideoBaseAdUnit.Parameters parameters = new VideoBaseAdUnit.Parameters();
+        VideoParameters parameters = new VideoParameters();
         parameters.setMimes(Collections.singletonList("video/mp4"));
         parameters.setProtocols(Collections.singletonList(Signals.Protocols.VAST_2_0));
         parameters.setPlaybackMethod(Collections.singletonList(Signals.PlaybackMethod.AutoPlaySoundOff));
