+++ Example/PrebidDemoKotlin/src/main/java/org/prebid/mobile/prebidkotlindemo/activities/ads/gam/original/GamOriginalApiDisplayBanner300x250Activity.kt
@@ -21,7 +21,7 @@ import com.google.android.gms.ads.AdSize
 import com.google.android.gms.ads.admanager.AdManagerAdRequest
 import com.google.android.gms.ads.admanager.AdManagerAdView
 import org.prebid.mobile.BannerAdUnit
-import org.prebid.mobile.BannerBaseAdUnit
+import org.prebid.mobile.BannerParameters
 import org.prebid.mobile.Signals
 import org.prebid.mobile.addendum.AdViewUtils
 import org.prebid.mobile.addendum.PbFindSizeError
@@ -52,7 +52,7 @@ class GamOriginalApiDisplayBanner300x250Activity : BaseAdActivity() {
         adUnit?.setAutoRefreshInterval(refreshTimeSeconds)
 
         // 2. Configure banner parameters
-        val parameters = BannerBaseAdUnit.Parameters()
+        val parameters = BannerParameters()
         parameters.api = listOf(Signals.Api.MRAID_3, Signals.Api.OMID_1)
         adUnit?.parameters = parameters
